Added move forward method to warehouses classes
