--- Team8923_2021/src/main/java/org/firstinspires/ftc/team8923_2021/BlueCarouselNoDetect.java
@@ -1,19 +1,0 @@
-package org.firstinspires.ftc.team8923_2021;
-
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-
-@Autonomous(name="BlueCarouselNoDetect")
-public class BlueCarouselNoDetect extends MasterAutonomous {
-    public void runOpMode() throws InterruptedException {
-        initAuto();
-        waitForStart();
-
-        while (opModeIsActive()) {
-            moveForward(16, 80, 10);
-            double referenceAngle = imu.getAngularOrientation().firstAngle; // Get a reference ange from the IMU for future movements using IMU
-            imuPivot(referenceAngle, -90, 30, 0.015, 3.0);
-            moveForward(50, 80, 10);
-            break;
-        }
-    }
-}
