+++ Team8923_2021/src/main/java/org/firstinspires/ftc/team8923_2021/BlueCarouselNoIntake.java
@@ -1,6 +1,4 @@
 package org.firstinspires.ftc.team8923_2021;
 
 public class BlueCarouselNoIntake {
-
-    //
 }
