--- Team8923_2021/src/main/java/org/firstinspires/ftc/team8923_2021/RedCarouselNoIntake.java
@@ -1,4 +1,0 @@
-package org.firstinspires.ftc.team8923_2021;
-
-public class RedCarouselNoIntake {
-}
